add initial oblique slicer
