+++ examples/oblique_slicing.py
@@ -8,28 +8,14 @@ import pygfx as gfx
 from qtpy.QtWidgets import QApplication, QGridLayout, QLabel, QWidget
 from wgpu.gui.qt import WgpuCanvas
 
+from cellier.slicer.transforms import AffineTransform
+
 
 def make_plane_mesh(
     point: np.ndarray, normal: np.ndarray, width: float
 ) -> Tuple[np.ndarray, np.ndarray]:
     """Make a mesh to visualize a plane."""
     pass
-
-
-def to_vec4(coordinates: np.ndarray) -> np.ndarray:
-    """Convert coordinates to vec4 to make compatible with an affine matrix."""
-    coordinates = np.atleast_2d(coordinates)
-
-    ndim = coordinates.shape[1]
-    if ndim == 3:
-        # add a 1 in the fourth dimension.
-        return np.pad(coordinates, pad_width=((0, 0), (0, 1)), constant_values=1)
-
-    elif coordinates.shape == 4:
-        return coordinates
-
-    else:
-        raise ValueError(f"Coordinates should be 3D or 4D, coordinates were {ndim}D")
 
 
 def compute_plane_coefficients(
@@ -168,65 +154,6 @@ class Oblique2DDataSliceRequest:
     spatial_dims: Tuple[int, int, int]
     slice_positive_extent: float
     slice_negative_extent: float
-
-
-@dataclass
-class AffineTransform:
-    """Affine transformation class."""
-
-    matrix: np.ndarray
-
-    def map_coordinates(self, coordinates: np.ndarray):
-        """Apply the transformation to coordinates."""
-        return np.dot(to_vec4(coordinates), self.matrix)[:, :3]
-
-    def imap_coordinates(self, coordinates: np.ndarray):
-        """Apply the inverse transformation to coordinates."""
-        return np.dot(to_vec4(coordinates), np.linalg.inv(self.matrix))[:, :3]
-
-    def map_normal_vector(self, normal_vector: np.ndarray):
-        """Apply the transform to a normal vector defining an orientation.
-
-        For example, this would be used to a plane normal.
-
-        https://www.scratchapixel.com/lessons/mathematics-physics-for-computer-graphics/geometry/transforming-normals.html
-
-        Parameters
-        ----------
-        normal_vector : np.ndarray
-            The normal vector(s) to be transformed.
-
-        Returns
-        -------
-        transformed_vector : np.ndarray
-            The transformed normal vectors as a unit vector.
-        """
-        normal_transform = np.linalg.inv(self.matrix).T
-        transformed_vector = np.matmul(to_vec4(normal_vector), normal_transform)[:, :3]
-
-        return transformed_vector / np.linalg.norm(transformed_vector, axis=1)
-
-    def imap_normal_vector(self, normal_vector: np.ndarray):
-        """Apply the inverse transform to a normal vector defining an orientation.
-
-        For example, this would be used to a plane normal.
-
-        https://www.scratchapixel.com/lessons/mathematics-physics-for-computer-graphics/geometry/transforming-normals.html
-
-        Parameters
-        ----------
-        normal_vector : np.ndarray
-            The normal vector(s) to be transformed.
-
-        Returns
-        -------
-        transformed_vector : np.ndarray
-            The transformed normal vectors as a unit vector.
-        """
-        normal_transform = self.matrix.T
-        transformed_vector = np.matmul(to_vec4(normal_vector), normal_transform)[:, :3]
-
-        return transformed_vector / np.linalg.norm(transformed_vector, axis=1)
 
 
 # make 4D points
