+++ src/cellier/slicer/__init__.py
@@ -1,1 +1,6 @@
 """Infrastructure to manage slicing of the data for viewing."""
+
+from cellier.slicer.slicer import SynchronousDataSlicer
+from cellier.slicer.world_slice import ObliqueWorldSlice
+
+__all__ = ["SynchronousDataSlicer", "ObliqueWorldSlice"]
