+++ tests/slicer/test_world_slice.py
@@ -1,0 +1,2 @@
+def test_axis_aligned_slice():
+    pass
